Crypto Prediction Service/Prediction Pipeline/requirements.txt
# Core ML/NLP
transformers==4.39.3
torch==2.2.2
tokenizers==0.15.2
sentencepiece==0.2.0
emoji==0.6.0

# Data handling
pandas==2.2.2
numpy==1.26.4
supabase>=2.3.5

# Utility
python-dotenv==1.0.1
tqdm==4.66.2
loguru==0.7.2
Flask>=2.2.0
gunicorn>=20.1.0
scikit-learn==1.3.2
joblib

# Async
aiohttp==3.9.5
